(RES): minor refactor
Move RustPat extensions to extensions
