+++ src/org/rust/lang/core/psi/impl/mixin/RustLambdaParamImplMixin.kt
@@ -4,7 +4,7 @@ import com.intellij.lang.ASTNode
 import org.rust.lang.core.psi.RustLambdaParam
 import org.rust.lang.core.psi.RustNamedElement
 import org.rust.lang.core.psi.impl.RustCompositeElementImpl
-import org.rust.lang.core.psi.util.boundElements
+import org.rust.lang.core.psi.boundElements
 
 public abstract class RustLambdaParamImplMixin(node: ASTNode)   : RustCompositeElementImpl(node)
                                                                 , RustLambdaParam {
