+++ src/org/rust/lang/core/psi/impl/mixin/RustLetDeclImplMixin.kt
@@ -4,7 +4,7 @@ import com.intellij.lang.ASTNode
 import org.rust.lang.core.psi.RustLetDecl
 import org.rust.lang.core.psi.RustNamedElement
 import org.rust.lang.core.psi.impl.RustCompositeElementImpl
-import org.rust.lang.core.psi.util.boundElements
+import org.rust.lang.core.psi.boundElements
 
 public abstract class RustLetDeclImplMixin(node: ASTNode)   : RustCompositeElementImpl(node)
                                                             , RustLetDecl {
