+++ src/org/rust/lang/core/psi/impl/mixin/RustPatFieldImplMixin.kt
@@ -4,7 +4,7 @@ import com.intellij.lang.ASTNode
 import org.rust.lang.core.psi.RustNamedElement
 import org.rust.lang.core.psi.RustPatField
 import org.rust.lang.core.psi.impl.RustNamedElementImpl
-import org.rust.lang.core.psi.util.boundElements
+import org.rust.lang.core.psi.boundElements
 
 public abstract class RustPatFieldImplMixin(node: ASTNode)  : RustNamedElementImpl(node)
                                                             , RustPatField {
